GFDataLoadingVC
- Added GFDataLoadingVC to restrict loading view/showing empty state view to UIViewControllers that inherit from GFDataLoadingVC
- Before, any view controller could call these functions and we may not want that behavior on all view controllers
